Add poweroff delay
Turning the key from "accessory" to "on" drops the ignition signal for a
moment, and the carputer used to shutdown only to turn back on again.
While this is a useful test for the power supply, just keep the thing
on. Also it's nice to have some moments to prevent a regretted shutdown.

Add $POWER_OFF_DELAY_SECS to specify shutdown delay